renderercommon: C++ compatibility for tr_noise

+++ code/renderercommon/tr_noise.c
@@ -59,11 +59,11 @@ float R_NoiseGet4f( float x, float y, float z, double t )
 	float back[4];
 	float fvalue, bvalue, value[2], finalvalue;
 
-	ix = ( int ) floor( x );
+	ix = ( int ) floor( (double)(x) );
 	fx = x - ix;
-	iy = ( int ) floor( y );
+	iy = ( int ) floor( (double)(y) );
 	fy = y - iy;
-	iz = ( int ) floor( z );
+	iz = ( int ) floor( (double)(z) );
 	fz = z - iz;
 	it = ( int ) floor( t );
 	ft = t - it;
